apps/neural-workbench-streamlit/requirements.txt
streamlit==1.29.0
pandas==2.1.4
plotly==5.17.0
numpy==1.24.3
supabase==2.3.4
python-dotenv==1.0.0